% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{A hello function}
\usage{
hello()
}
\description{
Prints \code{Hello World!}.
}
\details{
This function was written in R/hello.R, and all other functions in this
package were written in R Markdown documents.
}
\examples{
rlp::hello()
}
