% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo2-GEN.R
\docType{data}
\name{md}
\alias{md}
\title{A Markdown generator}
\format{An object of class \code{list} of length 6.}
\usage{
md
}
\value{
Each function element returns a character vector.
}
\description{
You can generate Markdown text using the function elements of this object.
}
\examples{
md$b("bold")
md$em("emphasize")
md$ul(head(letters))
}
\keyword{datasets}
